Refactor mcqa.
